DiceBCELossWithLogits: fix dtype mismatch for AMP (#179)

+++ mipcandy/common/optim/loss.py
@@ -40,7 +40,7 @@ class DiceBCELossWithLogits(nn.Module):
             if d not in (1, 2, 3):
                 raise ValueError(f"Expected labels to be 1D, 2D, or 3D, got {d} spatial dimensions")
             labels = convert_ids_to_logits(labels.int(), d, self.num_classes)
-        labels = labels.float()
+        labels = labels.to(dtype=masks.dtype)
         bce = nn.functional.binary_cross_entropy_with_logits(masks, labels)
         masks = masks.sigmoid()
         soft_dice = soft_dice_coefficient(masks, labels, smooth=self.smooth, include_background=self.include_background)
